design ebook segment API

+++ library-service/src/main/java/com/github/sujankumarmitra/libraryservice/v1/model/EBookSegment.java
@@ -12,7 +12,7 @@ public abstract class EBookSegment {
 
     public abstract String getBookId();
 
-    public abstract Long getIndex();
+    public abstract Integer getIndex();
 
     public abstract String getAssetId();
